+++ assets/components/ui/uiButton/uiAnchorButton.jsx
@@ -4,24 +4,22 @@
 
 import React from 'react';
 
-import { getClassName, defaultProps, type PropTypes } from './helpers';
+import DangerouslySetButtonOnAnyElement, { defaultProps, type PropTypes } from './dangerouslySetButtonOnAnyElement';
 import './uiButton.scss';
 
 // ----- Render ----- //
 
-const UiAnchorButton = ({
-  children, icon, type, disabled, appearance, iconSide, isStatic, modifierClasses, ...otherProps
-}: PropTypes) => (
-  <a
-    className={getClassName(appearance, iconSide, modifierClasses)}
-    {...otherProps}
-  >
-    <span className="component-ui-button__content">{children}</span>
-    {icon}
-  </a>
+type AllPropTypes = {
+  ...PropTypes,
+};
+
+const UiAnchorButton = (props: AllPropTypes) => (
+  <DangerouslySetButtonOnAnyElement
+    element="a"
+    {...props}
+  />
 );
 
-/* eslint-disable react/default-props-match-prop-types */
 UiAnchorButton.defaultProps = {
   ...defaultProps,
 };
