+++ assets/components/ui/uiButton/uiButton.scss
@@ -24,7 +24,7 @@
   }
 }
 
-.component-ui-button--static {
+.component-ui-button--placeholder {
   pointer-events: none;
 }
